requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
httpx>=0.27.0
pydantic-settings>=2.6.0
google-genai>=1.0.0
google-api-python-client>=2.150.0
google-auth>=2.36.0
google-auth-oauthlib>=1.2.0
python-dotenv>=1.0.0
